app/requirements.txt
flask
werkzeug
tensorflow>=2.10
Pillow 